+++ sources/OpenAL/Soft.Native/build-android-x86.sh
@@ -4,16 +4,17 @@ if [ ! -e ../../../eng/submodules/openal-soft/CMakeLists.txt ]; then
     git submodule update --init --recursive --depth 1 ../../../eng/submodules/openal-soft
 fi
 
-if [ ! -z "$GITHUB_ACTIONS" ]; then
+if [[ "$@" == *"--install-deps"* ]]; then
   # NDK already installed: https://github.com/actions/runner-images/blob/main/images/ubuntu/Ubuntu2404-Readme.md
   sdkmanager=( /usr/local/lib/android/sdk/cmdline-tools/*/bin/sdkmanager )
   ${sdkmanager[-1]} --install "platforms;android-21"
+  exit
 fi
 
 rm -rf build
 mkdir build
 cd build
-cmake ../../../../eng/submodules/openal-soft -DANDROID_ABI=x86 -DCMAKE_TOOLCHAIN_FILE="${ANDROID_NDK_HOME}/build/cmake/android.toolchain.cmake" -DANDROID_PLATFORM=21 -DCMAKE_BUILD_TYPE=Release -DALSOFT_UTILS=OFF -DALSOFT_EXAMPLES=OFF
+cmake ../../../../eng/submodules/openal-soft -DANDROID_ABI=x86 -DCMAKE_TOOLCHAIN_FILE="${ANDROID_NDK_HOME}/build/cmake/android.toolchain.cmake" -DANDROID_PLATFORM=21 -DCMAKE_BUILD_TYPE=Release -DALSOFT_UTILS=OFF -DALSOFT_EXAMPLES=OFF -DALSOFT_EMBED_HRTF_DATA=TRUE -DALSOFT_REQUIRE_OPENSL=ON
 cmake --build . --parallel
 cd ..
 mkdir -p runtimes/android-x86/native
